sprand: intersperse repeated offsets with new ones
Because the exact number of erase blocks does not need to be used for
sprand to work, it is helpful to sprinkle repeated offsets among the new
offsets. The previous approach began with a stream of unique offsets and
followed that with offsets chosen from the ones already provided. The
result is a stream of mostly valid offsets at the end of each erase
block. If the actual number of erase blocks is not known, it is better
to sprinkle invalidating offsets among the new offsets.

This patch changes the strategy for providing offsets. The first offset
for each erase block is always a new one. Thereafter, an appropriately
biased coin is flipped to determine whether to provide a new offset or
to invalidate a previously written offset by repeating one that was
already used. The coin flip is produced by a new LFSR instance for each
erase block to ensure that we do not run out of new valid offsets. New
offsets written early in the erase block are more likely to appear among
the repeated offsets because these offsets are chosen from the ones
already written to the erase block.