sprand: use weighted average of number of valid pages
Instead of using the final valid page count that pushes the sum of valid
page fractions over 1, use the weighted average of valid page counts
where the weights are the fractions associated with each valid page
count.

Also use a new random seed for each within-page LFSR. On occasion this
LFSR reset fails and we need to re-use the previous random seed.